Add files via upload
Dynamic Information Network concepts, and Flux Explorations. 

The first way we consciously experience the universe and reality is through our body. This bias is intense, as it directs the data (stimulus) we seek and directs how we perceive data in general. 

When A.I seek to describe or develop a perspective on reality based on their frame of reference. Naturally, information theory often takes precedence.